requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pillow
streamlit
requests
numpy
tqdm
pandas
